refactor: Create models folder and move keys into it
