+++ packages/neuron-wallet/src/addresses/dao.ts
@@ -1,6 +1,6 @@
 import { Not } from 'typeorm'
 import AddressEntity, { AddressVersion } from './entities/address'
-import { AddressType } from '../keys/address'
+import { AddressType } from '../models/keys/address'
 import { getConnection } from './ormconfig'
 import TransactionsService from '../services/transactions'
 import CellsService from '../services/cells'
