+++ packages/neuron-wallet/src/controllers/wallets/index.ts
@@ -1,10 +1,10 @@
 import fs from 'fs'
 import AppController from '../app'
 import WalletsService, { Wallet, WalletProperties, FileKeystoreWallet } from '../../services/wallets'
-import Keystore from '../../keys/keystore'
-import Keychain from '../../keys/keychain'
-import { validateMnemonic, mnemonicToSeedSync } from '../../keys/mnemonic'
-import { AccountExtendedPublicKey, ExtendedPrivateKey } from '../../keys/key'
+import Keystore from '../../models/keys/keystore'
+import Keychain from '../../models/keys/keychain'
+import { validateMnemonic, mnemonicToSeedSync } from '../../models/keys/mnemonic'
+import { AccountExtendedPublicKey, ExtendedPrivateKey } from '../../models/keys/key'
 import { Controller as ControllerDecorator, CatchControllerError } from '../../decorators'
 import { ResponseCode, Channel } from '../../utils/const'
 import {
