+++ packages/neuron-wallet/src/models/keys/keystore.ts
@@ -2,7 +2,7 @@ import crypto from 'crypto'
 import SHA3 from 'sha3'
 import { v4 as uuid } from 'uuid'
 
-import { UnsupportedCipher, IncorrectPassword } from '../exceptions'
+import { UnsupportedCipher, IncorrectPassword } from '../../exceptions'
 import { ExtendedPrivateKey } from './key'
 
 const CIPHER = 'aes-128-ctr'
