+++ packages/neuron-wallet/src/services/wallets.ts
@@ -2,8 +2,8 @@ import { v4 as uuid } from 'uuid'
 import { fromEvent } from 'rxjs'
 import { debounceTime } from 'rxjs/operators'
 import TransactionsService from './transactions'
-import { AccountExtendedPublicKey, PathAndPrivateKey } from '../keys/key'
-import Keystore from '../keys/keystore'
+import { AccountExtendedPublicKey, PathAndPrivateKey } from '../models/keys/key'
+import Keystore from '../models/keys/keystore'
 import Store from '../utils/store'
 import NodeService from './node'
 import FileService from './file'
@@ -16,7 +16,7 @@ import Blake2b from '../utils/blake2b'
 import { CurrentWalletNotSet, WalletNotFound, IsRequired, UsedName } from '../exceptions'
 import AddressService from './addresses'
 import { Address as AddressInterface } from '../addresses/dao'
-import Keychain from '../keys/keychain'
+import Keychain from '../models/keys/keychain'
 import { updateApplicationMenu } from '../utils/application-menu'
 import AddressesUsedSubject from '../subjects/addresses-used-subject'
