+++ packages/neuron-wallet/tests/models/keys/key.test.ts
@@ -1,5 +1,10 @@
-import { ExtendedPublicKey, AccountExtendedPublicKey, ExtendedPrivateKey, generateMnemonic } from '../../src/keys/key'
-import { AddressType } from '../../src/keys/address'
+import {
+  ExtendedPublicKey,
+  AccountExtendedPublicKey,
+  ExtendedPrivateKey,
+  generateMnemonic,
+} from '../../../src/models/keys/key'
+import { AddressType } from '../../../src/models/keys/address'
 
 const fixture = {
   privateKey: 'e8f32e723decf4051aefac8e2c93c9c5b214313817cdb01a1494b917c8436b35',
