+++ packages/neuron-wallet/tests/services/wallets.test.ts
@@ -2,10 +2,10 @@ import '../../src/controllers/wallets'
 import '../../src/controllers/app'
 import WalletService, { WalletProperties } from '../../src/services/wallets'
 import { Witness } from '../../src/app-types/types'
-import Keystore from '../../src/keys/keystore'
-import Keychain from '../../src/keys/keychain'
-import { mnemonicToSeedSync } from '../../src/keys/mnemonic'
-import { ExtendedPrivateKey, AccountExtendedPublicKey } from '../../src/keys/key'
+import Keystore from '../../src/models/keys/keystore'
+import Keychain from '../../src/models/keys/keychain'
+import { mnemonicToSeedSync } from '../../src/models/keys/mnemonic'
+import { ExtendedPrivateKey, AccountExtendedPublicKey } from '../../src/models/keys/key'
 
 describe('wallet service', () => {
   let walletService: WalletService
